makefile for compiling my brand new file
