+++ src/train_manager/train_manager.c
@@ -5,9 +5,9 @@ void trainManagerThread() {}
 
 int main(int argc, char **argv) {
   // create 2 threads
-  Train_t train1;
+  struct Train_t train1;
   train1.trainId = 1;
-  train1.pls;
+  train1.plc;
   train1.resManager;
   train1.path = initPath("route/train1.route");
